Add components to base layout
Contents and page header. The data cascade may get confusing here, but I'm trying to make it so that we only need the base layout. It may be necessary to have a separate layout for each page, which isn't a disaster. But we'll cross that bridge when we come to it.